Update g_game.c
g_game.c:289:16: warning: comparison of integer expressions of different signedness: 'int' and 'long long unsigned int' [-Wsign-compare]
  289 |     for (i=0; i<arrlen(weapon_order_table); ++i)
      |                ^
g_game.c: In function 'G_BuildTiccmd':
g_game.c:504:20: warning: comparison of integer expressions of different signedness: 'int' and 'long long unsigned int' [-Wsign-compare]
  504 |         for (i=0; i<arrlen(weapon_keys); ++i)
      |                    ^

+++ src/doom/g_game.c
@@ -286,7 +286,7 @@ static int G_NextWeapon(int direction)
         weapon = players[consoleplayer].pendingweapon;
     }
 
-    for (i=0; i<arrlen(weapon_order_table); ++i)
+    for (i=0; (size_t)i<arrlen(weapon_order_table); ++i)
     {
         if (weapon_order_table[i].weapon == weapon)
         {
@@ -501,7 +501,7 @@ void G_BuildTiccmd (ticcmd_t* cmd, int maketic)
     {
         // Check weapon keys.
 
-        for (i=0; i<arrlen(weapon_keys); ++i)
+        for (i=0; (size_t)i<arrlen(weapon_keys); ++i)
         {
             int key = *weapon_keys[i];
